# Make `unary` and `binary` faster