@alamb 2024-09-09
Thank you @AdamGS  -- I plan to review this carefully over the next few days

cc @tustvold @jhorstmann @viirya